complete task

+++ 0x0B-malloc_free/1-strdup.c
@@ -11,7 +11,7 @@
 char *_strdup(char *str)
 {
 	char *ptr;
-	int i;
+	long unsigned int i;
 
 	if (strlen(str) == 0 || str == NULL)
 	{
